+++ app/src/main/java/com/geocipher/app/interfaces/ILocationManager.kt
@@ -6,4 +6,5 @@ interface ILocationManager {
     fun getUserLocation(activity: Activity, listener: LocationUpdateListener)
     fun requestLocationPermission(activity: Activity)
     fun checkLocationPermission() : Boolean
+    fun checkGPSEnabled(): Boolean
 }
