+++ app/src/main/java/com/geocipher/app/interfaces/IMessageRepository.kt
@@ -1,5 +1,6 @@
 package com.geocipher.app.interfaces
 
 interface IMessageRepository {
-    fun addMessage(message: String, key: String, latitude: Double, longitude: Double)
+    fun addMessage(encryptedMessage: String, key: String, latitude: Double, longitude: Double)
+    fun retrieveMessage(key: String, latitude: Double, longitude: Double)
 }
